// src/app/models/showroom.model.ts
export interface Showroom {
    id?: number; // Optional because it is generated by the backend
    name: string;
    commercialRegistrationNumber: string;
    managerName?: string; // Optional because it is not required
    contactNumber: string;
    address?: string; // Optional because it is not required
    active: boolean;
    cars?: Car[]; 
    
  }
  
  export interface Car {
    id?: number;
    name: string;
    model: string;
    price: number;
    showroomId?: number; 
  }